src/script/requirements.txt
pandas
bs4
requests
rich
pydantic
pillow
cloudpickle